backend/requirements.txt
scrapy>=2.11.0
beautifulsoup4>=4.12.0
openai>=1.0.0,<2.0.0
selenium>=4.15.0
python-dotenv>=1.0.0
requests>=2.31.0
python-multipart>=0.0.6
python-docx>=1.0.1
Pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
html5lib>=1.1
cssutils>=2.7.1
webcolors>=1.13
numpy>=1.24.3
urllib3>=2.0.7
fastapi>=0.104.0
uvicorn>=0.24.0
supabase==2.3.4
webdriver-manager>=4.0.1
reportlab>=4.0.0
markdown>=3.5.0 
stripe==7.7.0
httpx>=0.25.2 